refactor(proto): separate fetch handler, resources, hostscript
refactor(proto): move tests to fetch handler

feat(proto): separate sf runner

feat(proto): doc comment added + sf_runner removed

feat(jstzd): build image with sequencer

refactor(proto): separate fetch handler, resources, hostscript

refactor(proto): move tests to fetch handler

feat(proto): separate sf runner

feat(proto): doc comment added + sf_runner removed

feat(proto): remove pub
