+++ crates/jstz_proto/src/runtime/v2/fetch/host_script.rs
@@ -1,0 +1,8 @@
+pub struct HostScript;
+
+impl HostScript {
+    #[allow(dead_code)]
+    pub async fn handle_balance() {
+        todo!()
+    }
+}
